+++ src/renderer/App.tsx
@@ -10,6 +10,8 @@ import UIContext from "./UIContext";
 import UIReducer from "./UIReducer";
 import TitleBar from "./TitleBar";
 import SidebarKnob from "./SidebarKnob";
+import PDFReader from "./PDFReader";
+import * as ipc from "./ipc";
 
 const Container = styled.div`
   width: 100%;
@@ -53,12 +55,46 @@ const App = () => {
     setShowKnob(true && !uiState.showSidebar);
   }, [uiState]);
 
+  const preventDefaults = (e: React.MouseEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+  };
+
+  const appendFile = (e: React.DragEvent) => {
+    preventDefaults(e);
+
+    Array.from(e.dataTransfer.files).forEach((file) => {
+      if (file.type.startsWith("image/") || file.type.startsWith("video/")) {
+        appDispatch({
+          type: "APPEND_PAGE",
+          src: `file://${file.path}`,
+          contentType: file.type,
+        });
+      } else if (file.type === "application/pdf") {
+        (async () => {
+          const pdf = await PDFReader.loadURL(`file://${file.path}`);
+          pdf.eachPage(async (page) => {
+            const png = await page.getPNG({ width: 1920 });
+            const p = await ipc.saveTempFile(png, ".png");
+            appDispatch({
+              type: "APPEND_PAGE",
+              src: `file://${p}`,
+              contentType: "image/png",
+            });
+          });
+        })();
+      }
+    });
+  };
+
   return (
     <AppContext.Provider value={{ state: appState, dispatch: appDispatch }}>
       <UIContext.Provider value={{ state: uiState, dispatch: uiDispatch }}>
         <Container
           onMouseEnter={showSidebarKnob}
           onMouseLeave={hideSidebarKnob}
+          onDragOver={preventDefaults}
+          onDrop={appendFile}
         >
           <VerticalSplit
             left={<PageList />}
@@ -68,8 +104,8 @@ const App = () => {
             minLeftWidth={80}
             onResized={handleSidebarResized}
           />
-          <TitleBar shown={showTitleBar} text={"Sanscadre"} />
           <SidebarKnob shown={showKnob} onClick={handleKnobClicked} />
+          <TitleBar shown={showTitleBar} text={"Sanscadre"} />
         </Container>
         <KeyHandler target={window.document.documentElement} />
       </UIContext.Provider>
