+++ src/renderer/Screen/index.ts
@@ -1,0 +1,3 @@
+import Screen from "./Screen";
+
+export default Screen;
